Added in correct AK and NAK sequence for acc rd
Huge debugging problem.  Actual root cause was the fact that the
multiple register read requires AK's after every received byte and then
the last byte requires a NAK.  The sequence had all NAKS before which
would allow me to still read the data coming back, but the interrupt was
not being cleared because it thought that we were still reading
registers.